third_party/blink/web_tests/virtual/select-parser-relaxation/external/wpt/html/semantics/forms/the-select-element/stylable-select/selectedoptionelement-attr.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] The selectedoptionelement attribute should set up an association between a select and a selectedoption.
  assert_equals: Setting the selectedoptionelement attribute via setAttribute should synchronously assign the contents of <selectedoption>. expected "one<span>span</span>" but got ""
Harness: the test ran to completion.
